+++ source/icu.net.tests/SetUpFixture.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2017 SIL International
+// Copyright (c) 2017-2022 SIL International
 // This software is licensed under the MIT license (http://opensource.org/licenses/MIT)
 
 using NUnit.Framework;
@@ -16,6 +16,11 @@ namespace Icu.Tests
 		public void RunBeforeAnyTests()
 		{
 			Wrapper.Init();
+
+			// Limit maximum version to the version we install, otherwise some tests might fail
+			// if we find a higher version on the PATH.
+			Wrapper.ConfineIcuVersions(Wrapper.MinSupportedIcuVersion,
+				NativeMethodsTests.MaxInstalledIcuLibraryVersion);
 		}
 
 #if NUNIT2